pyproject.toml
[project]
name = "amelia"
version = "0.0.1"
requires-python = ">=3.10.18"
dependencies = [
    "aiohttp==3.7.4.post0",
    "annotated-types==0.5.0",
    "async-timeout==3.0.1",
    "asyncpg==0.28.0",
    "asyncpg-stubs==0.28.0",
    "attrs==21.2.0",
    "certifi==2021.10.8",
    "chardet==4.0.0",
    "colorama==0.4.4",
    "discord-py==2.3.2",
    "exceptiongroup==1.1.3",
    "faker==19.3.1",
    "fuzzywuzzy==0.18.0",
    "idna==3.3",
    "iniconfig==1.1.1",
    "multidict==6.0.2",
    "nodeenv==1.8.0",
    "packaging==20.9",
    "pdf2image==1.16.0",
    "pillow==9.2.0",
    "pluggy==0.13.1",
    "polyfactory==2.8.0",
    "psycopg2-binary==2.9.1",
    "py==1.10.0",
    "pydantic==2.3.0",
    "pydantic-core==2.6.3",
    "pyparsing==2.4.7",
    "pyright==1.1.327",
    "pytest==7.4.2",
    "pytest-asyncio==0.20.3",
    "pytest-dotenv==0.5.2",
    "python-dateutil==2.8.2",
    "python-dotenv==0.19.0",
    "python-levenshtein==0.12.2",
    "ruff==0.0.289",
    "sentry-sdk==1.3.1",
    "six==1.16.0",
    "sqlparse==0.4.1",
    "tabulate==0.8.9",
    "toml==0.10.2",
    "tomli==2.0.1",
    "typing-extensions==4.7.1",
    "urllib3==1.26.6",
    "yarl==1.6.3",
    "yoyo-migrations==7.3.2",
]

[tool.ruff]
# Allow lines to be as long as 120 characters.
line-length = 120

[tool.uv]
package = false
